Separate command-related logic into its own file
